refractored modular_robot module to make it more general
