+++ examples/robot_brain_cmaes/main.py
@@ -1,7 +1,7 @@
 """
 Set up an experiment that optimizes the brain of a given robot body using CMA-ES.
 
-As the body is static, the genotype of the brain will be a fixed length real-valued vector.
+As the body is static, the genotype of the brain will be a fixed length real-valued custom_vector.
 
 Before starting this tutorial, it is useful to look at the 'experiment_setup' and 'evaluate_multiple_isolated_robots' examples.
 It is also nice to understand the concept of a cpg brain, although not really needed.
