+++ examples/simulate_single_robot/main.py
@@ -12,8 +12,9 @@ from revolve2.ci_group import terrains
 from revolve2.ci_group.logging import setup_logging
 from revolve2.ci_group.rng import make_rng
 from revolve2.ci_group.simulation import create_batch_single_robot_standard
-from revolve2.modular_robot import ActiveHinge, Body, Brick, ModularRobot, RightAngles
-from revolve2.modular_robot.brains import BrainCpgNetworkNeighborRandom
+from revolve2.modular_robot._common import ActiveHinge, Body, Brick, RightAngles
+from revolve2.modular_robot.v1 import ModularRobotV1 as ModularRobot
+from revolve2.modular_robot._common.brains import BrainCpgNetworkNeighborRandom
 from revolve2.simulators.mujoco import LocalRunner
