+++ modular_robot/revolve2/modular_robot/brains/_brain_cpg_network_neighbor.py
@@ -3,7 +3,7 @@ from abc import ABC, abstractmethod
 
 from revolve2.actor_controller import ActorController
 from revolve2.actor_controllers.cpg import CpgActorController as ControllerCpg
-from revolve2.modular_robot import ActiveHinge, Body, Brain
+from revolve2.modular_robot._common import ActiveHinge, Body, Brain
 
 from ._make_cpg_network_structure_neighbor import (
     active_hinges_to_cpg_network_structure_neighbor,
