+++ modular_robot/revolve2/modular_robot/brains/_brain_cpg_network_neighbor_random.py
@@ -1,5 +1,5 @@
 import numpy as np
-from revolve2.modular_robot import ActiveHinge, Body
+from revolve2.modular_robot._common import ActiveHinge, Body
 
 from ._brain_cpg_network_neighbor import BrainCpgNetworkNeighbor
