+++ modular_robot/revolve2/modular_robot/v1/__init__.py
@@ -1,0 +1,5 @@
+"""Classes and functions for V1 Hardware."""
+
+from ._modular_robot_v1 import ModularRobotV1
+
+__all__ = ["ModularRobotV1"]
