+++ modular_robot/revolve2/modular_robot/v2/__init__.py
@@ -1,0 +1,5 @@
+"""Classes and functions for V2 Hardware."""
+from ._bodyv2 import BodyV2
+from ._modular_robotv2 import ModularRobotV2
+
+__all__ = ["BodyV2", "ModularRobotV2"]
