+++ simulators/mujoco/mypy.ini
@@ -16,3 +16,6 @@ ignore_missing_imports = True
 
 [mypy-cv2.*]
 ignore_missing_imports = True
+
+[mypy-cv2.*]
+ignore_missing_imports = True
